Use faster intervals impl
