+++ benches/lib.rs
@@ -1,6 +1,10 @@
+#[cfg(feature = "ripe-geo-embedded")]
 use criterion::criterion_main;
 
-#[cfg(feature = "ripe-geo-embedded")]
-mod interval_tree;
+mod intervals;
 
-criterion_main!(interval_tree::benches_interval_tree);
+#[cfg(feature = "ripe-geo-embedded")]
+criterion_main!(intervals::benches_interval_tree);
+
+#[cfg(not(feature = "ripe-geo-embedded"))]
+fn main() {}
